Spring Cleaning
The mess that was the Interface folder is now much better. I still
retain the old files so not to worry. If any problems arise that you
notice because of this let me know ASAP.

P.S I guarantee the saved checksheets will not work or look right after
this so they need to be fixed by changing function names/urls